requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pymongo==4.6.1
motor==3.3.2
langchain==0.1.6
langchain-text-splitters==0.0.1
langchain-community==0.0.16
sentence-transformers==2.4.0
python-dotenv==1.0.1
pydantic==2.6.3
google-generativeai==0.3.2
transformers==4.38.2
torch==2.2.1